+++ CHANGES.md
@@ -1,10 +1,10 @@
 # Release History
 
-## 1.36.2 (2024-mm-dd)
+## 1.37.0 (2024-04-22)
 
+- **ENH: Add AWS profile feature to s3 module**
 - FIX: Fix `files.extract_file` when there are only files in the root of the zip archive ([#14](https://github.com/sertit/sertit-utils/pull/14))
 - FIX: FIX metadata handling with `rasters_rio.read` when reading with indexes
-- **ENH: Add AWS profile feature to s3 module**
 
 ## 1.36.1 (2024-03-26)